separate types
