+++ packages/types/src/index.ts
@@ -2,7 +2,6 @@ export * from './types/database-interface';
 export * from './types/connection-informations';
 export * from './types/tokens';
 export * from './types/token-record';
-export * from './types/session';
 export * from './types/user';
 export * from './types/create-user';
 export * from './types/email-record';
@@ -12,3 +11,6 @@ export * from './types/login-user-identity';
 export * from './types/hook-listener';
 export * from './types/authentication-service';
 export * from './types/hash-algorithm';
+export * from './types/session/session';
+export * from './types/session/database-interface';
+export * from './types/services/password/database-interface';
